Move several structs needed for `StacksTransaction`
